refactor: adapt to Disman::Mode refresh rate type change
The refresh rate of Disman's Mode class is changed to double type. Adapt to
this change.

+++ kded/output.cpp
@@ -68,7 +68,7 @@ void Output::readInGlobalPartFromInfo(Disman::OutputPtr output, const QVariantMa
         if (mode->size() != size) {
             continue;
         }
-        if (!qFuzzyCompare(mode->refreshRate(), modeInfo[QStringLiteral("refresh")].toFloat())) {
+        if (!qFuzzyCompare(mode->refreshRate(), modeInfo[QStringLiteral("refresh")].toDouble())) {
             continue;
         }
